+++ CHANGELOG.md
@@ -5,8 +5,7 @@
 * Two new functions to customize tick labels on any existing plot:
     * `wrap_tick_labels`: add text wrapping
     * `add_sample_size_to_labels`: add group sample sizes with a `(n=N)` suffix
-
-
+* Make writing PDF figures a deterministic process and make the PDF text editable.
 
 ## 0.5.0 (2022-01-10)